Use LiveData with ViewModel 5. Attach observer to the LiveData object
In this task you set up an observer in the app component, GameFragment. The observer you will add observes the changes to the app's data currentScrambledWord. LiveData is lifecycle-aware, meaning it only updates observers that are in an active lifecycle state. So the observer in the GameFragment will only be notified when the GameFragment is in STARTED or RESUMED states.