Use LiveData with ViewModel 7. Use LiveData with data binding
In the previous tasks, your app listens to the data changes in the code. Similarly, apps can listen to the data changes from the layout. With Data Binding, when an observable LiveData value changes, the UI elements in the layout it's bound to are also notified, and the UI can be updated from within the layout.