Use LiveData with ViewModel 9. Use binding expressions
Binding expressions are written within the layout in the attribute properties (such as android:text) referencing the layout properties. Layout properties are declared at the top of the data binding layout file, via the <variable> tag. When any of the dependent variables change, the ‘DB Library' will run your binding expressions (and thus updates the views). This change-detection is a great optimization which you get for free, when you use a Data Binding Library.